✨ print cpio file content

+++ impl-c/include/string.h
@@ -6,4 +6,4 @@ int strcmp(const char *, const char *);
 int strncmp(const char *str1, const char *str2, size_t n);
 int strlen(const char *str);
 
-char* itoa(int val, int base);
+char* itoa(int64_t val, int base);
